+++ packages/oid-federation/src/index.ts
@@ -2,4 +2,4 @@ export * from "./entityConfiguration";
 export * from "./entityStatement";
 export * from "./jwk";
 export * from "./metadata";
-export * from "./utils";
+export type * from "./utils";