+++ packages/oid-federation/src/utils/types.ts
@@ -1,0 +1,6 @@
+import { JsonWebKey } from "../jwk";
+
+export type SignCallback = (options: {
+  jwk: JsonWebKey;
+  toBeSigned: Uint8Array;
+}) => Promise<Uint8Array>;