+++ packages/utils/src/index.ts
@@ -3,10 +3,12 @@ export * from "./errors";
 export * from "./fetcher";
 
 export {
+  type Fetch,
   JsonParseError,
   ValidationError,
   addSecondsToDate,
   dateToSeconds,
   decodeUtf8String,
   encodeToBase64Url,
+  parseWithErrorHandling,
 } from "@openid4vc/utils";